+++ model/utils.go
@@ -1,13 +1,13 @@
 package model
 
 import (
-	"database/sql"
 	"time"
 
 	"github.com/labstack/echo"
 	"gopkg.in/guregu/null.v3"
 )
 
+// NullTimeToString null許容の時間のStringへの変換
 func NullTimeToString(t null.Time) string {
 	if t.Valid {
 		return t.Time.Format(time.RFC3339)
@@ -16,14 +16,7 @@ func NullTimeToString(t null.Time) string {
 	return "null"
 }
 
-func NullStringConvert(str sql.NullString) string {
-	if str.Valid {
-		return str.String
-	}
-
-	return "NULL"
-}
-
+// GetUserID ユーザーIDの取得
 func GetUserID(c echo.Context) string {
 	res := c.Request().Header.Get("X-Showcase-User")
 	// test用
