+++ router/responses.go
@@ -34,17 +34,17 @@ func PostResponse(c echo.Context) error {
 
 	//パターンマッチ
 	for _, body := range req.Body {
-		validation, err := model.GetValidations(body.QuestionID)
+		validation, err := model.GetValidation(body.QuestionID)
 		if err != nil {
 			return echo.NewHTTPError(http.StatusInternalServerError, err)
 		}
 		switch body.QuestionType {
 		case "LinearScale":
-			label, err := model.GetScaleLabels(body.QuestionID)
+			label, err := model.GetScaleLabel(body.QuestionID)
 			if err != nil {
 				return echo.NewHTTPError(http.StatusInternalServerError, err)
 			}
-			if err := model.CheckScaleLabels(label, body.Body.ValueOrZero()); err != nil {
+			if err := model.CheckScaleLabel(label, body.Body.ValueOrZero()); err != nil {
 				return echo.NewHTTPError(http.StatusBadRequest, err)
 			}
 		case "Number":
@@ -92,56 +92,6 @@ func PostResponse(c echo.Context) error {
 	})
 }
 
-// GetMyResponses GET /users/me/responses
-func GetMyResponses(c echo.Context) error {
-	userID := model.GetUserID(c)
-	myResponses, err := model.GetRespondentInfos(c, userID)
-	if err != nil {
-		return err
-	}
-
-	return c.JSON(http.StatusOK, myResponses)
-}
-
-// GetMyResponsesByID GET /users/me/responses/:questionnaireID
-func GetMyResponsesByID(c echo.Context) error {
-	userID := model.GetUserID(c)
-	questionnaireID, err := strconv.Atoi(c.Param("questionnaireID"))
-	if err != nil {
-		c.Logger().Error(err)
-		return echo.NewHTTPError(http.StatusBadRequest)
-	}
-
-	myresponses, err := model.GetRespondentInfos(c, userID, questionnaireID)
-	if err != nil {
-		return err
-	}
-
-	return c.JSON(http.StatusOK, myresponses)
-}
-
-// GetResponsesByID GET /results/:questionnaireID
-func GetResponsesByID(c echo.Context) error {
-	sort := c.QueryParam("sort")
-	questionnaireID, err := strconv.Atoi(c.Param("questionnaireID"))
-	if err != nil {
-		c.Logger().Error(err)
-		return echo.NewHTTPError(http.StatusBadRequest)
-	}
-
-	// アンケートの回答を確認する権限が無ければエラーを返す
-	if err := checkResponseConfirmable(c, questionnaireID); err != nil {
-		return err
-	}
-
-	respondentDetails, err := model.GetRespondentDetails(c, questionnaireID, sort)
-	if err != nil {
-		return err
-	}
-
-	return c.JSON(http.StatusOK, respondentDetails)
-}
-
 // GetResponse GET /responses/:id
 func GetResponse(c echo.Context) error {
 	responseID, err := strconv.Atoi(c.Param("id"))
@@ -185,17 +135,17 @@ func EditResponse(c echo.Context) error {
 
 	//パターンマッチ
 	for _, body := range req.Body {
-		validation, err := model.GetValidations(body.QuestionID)
+		validation, err := model.GetValidation(body.QuestionID)
 		if err != nil {
 			return echo.NewHTTPError(http.StatusInternalServerError, err)
 		}
 		switch body.QuestionType {
 		case "LinearScale":
-			label, err := model.GetScaleLabels(body.QuestionID)
+			label, err := model.GetScaleLabel(body.QuestionID)
 			if err != nil {
 				return echo.NewHTTPError(http.StatusInternalServerError, err)
 			}
-			if err := model.CheckScaleLabels(label, body.Body.ValueOrZero()); err != nil {
+			if err := model.CheckScaleLabel(label, body.Body.ValueOrZero()); err != nil {
 				return echo.NewHTTPError(http.StatusBadRequest, err)
 			}
 		case "Number":
@@ -213,10 +163,6 @@ func EditResponse(c echo.Context) error {
 				return echo.NewHTTPError(http.StatusInternalServerError, err)
 			}
 		}
-	}
-
-	if err := model.UpdateRespondents(c, req.ID, responseID); err != nil {
-		return err
 	}
 
 	//全消し&追加(レコード数爆発しそう)
@@ -254,38 +200,4 @@ func DeleteResponse(c echo.Context) error {
 	}
 
 	return c.NoContent(http.StatusOK)
-}
-
-// アンケートの回答を確認できるか
-func checkResponseConfirmable(c echo.Context, questionnaireID int) error {
-	resSharedTo, err := model.GetResShared(c, questionnaireID)
-	if err != nil {
-		return err
-	}
-
-	switch resSharedTo {
-	case "administrators":
-		AmAdmin, err := model.CheckAdmin(c, questionnaireID)
-		if err != nil {
-			return err
-		}
-		if !AmAdmin {
-			return echo.NewHTTPError(http.StatusUnauthorized)
-		}
-	case "respondents":
-		AmAdmin, err := model.CheckAdmin(c, questionnaireID)
-		if err != nil {
-			return err
-		}
-		if !AmAdmin {
-			isRespondent, err := model.IsRespondent(c, questionnaireID)
-			if err != nil {
-				return err
-			}
-			if !isRespondent {
-				return echo.NewHTTPError(http.StatusUnauthorized, errors.New("only admins and respondents can see this responses"))
-			}
-		}
-	}
-	return nil
 }
